replace int with float

+++ main.cpp
@@ -1,24 +1,24 @@
 #include <iostream>
 using namespace std;
 
-void add (int num1, int num2) {
+void add (float num1, float num2) {
    cout << num1 << " + " << num2 << " = " << (num1 + num2) << endl;
 }
 
-void subtract(int num1, int num2) {
+void subtract(float num1, float num2) {
     cout << num1 << " - " << num2 << " = " << (num1 - num2) << endl;
 }
 
-void multiply(int num1, int num2) {
+void multiply(float num1, float num2) {
     cout << num1 << " * " << num2 << " = " << (num1 * num2) << endl;
 }
 
-void divide(int num1, int num2){
+void divide(float num1, float num2){
     cout << num1 << " / " << num2 << " = " << (num1 / num2) << endl;
 }
 
 int get_num(bool first) {
-    int num{};
+    float num{};
     if (first) {
         cout << "enter first number=> ";
         cin >> num;
@@ -38,8 +38,8 @@ void get_op () {
         << "4. Division" << endl;
     cout << "=> ";
     cin >> opt;
-    int num1 = get_num(true);
-    int num2 = get_num(false);
+    float num1 = get_num(true);
+    float num2 = get_num(false);
     switch (opt) {
         case 1:{
                    add(num1,num2);
